==> Submit/1.Setups.ipynb <==
# 1. Setups

__Package Install__

!pip install --upgrade pip

!cp ./files/requirements.txt ./

import time
tried=0
while True:
    print("tries:",tried)
    !pip install -r requirements.txt
    time.sleep(2.5)
    tried+=1
    if tried>5:break
        
# could not install packages due to an environmenterror errno 16 device or resource busy 에러가 가끔 발생. 
# 이를 위해 추가적으로 몇번 더 시도

__Font Install__

!cp malgun.ttf /opt/app-root/src/

!cp /opt/app-root/src/malgun.ttf /opt/app-root/lib/python3.6/site-packages/matplotlib/mpl-data/fonts/ttf/

!rm -rf /opt/app-root/src/.cache/matplotlib/*

from IPython.core.display import HTML
HTML("<script>Jupyter.notebook.kernel.restart()</script>")

#Font
import matplotlib
from matplotlib import font_manager
import matplotlib.pyplot as plt
import warnings

warnings.filterwarnings("ignore")
matplotlib.rcParams['axes.unicode_minus']=False

font_fname = '/opt/app-root/lib/python3.6/site-packages/matplotlib/mpl-data/fonts/ttf/malgun.ttf'
font_family = font_manager.FontProperties(fname=font_fname).get_name()
plt.rcParams["font.family"] = font_family
print("Font set to :",plt.rcParams["font.family"])
