==> Submit/2.Data Load& Preprocessing.ipynb <==
# 2. Data Load& Preprocessing

import os
import pathlib
import pandas as pd
import numpy as np
import geopandas as gpd
from geoband.API import GetCompasData
import json

#Font
import matplotlib
from matplotlib import font_manager
import matplotlib.pyplot as plt
import warnings

warnings.filterwarnings("ignore")
matplotlib.rcParams['axes.unicode_minus']=False

font_fname = '/opt/app-root/lib/python3.6/site-packages/matplotlib/mpl-data/fonts/ttf/malgun.ttf'
font_family = font_manager.FontProperties(fname=font_fname).get_name()
plt.rcParams["font.family"] = font_family

path = pathlib.Path('./data')

> ## __Get data__ 

if not path.is_dir():
    path.mkdir()
    
    GetCompasData('SBJ_2012_001', '1', path.joinpath('1.오산시_주정차단속(2018~2020).csv'))
    GetCompasData('SBJ_2012_001', '2', path.joinpath('2.오산시_어린이교통사고_격자.geojson'))
    GetCompasData('SBJ_2012_001', '3', path.joinpath('3.오산시_차량등록현황_격자.geojson'))
    GetCompasData('SBJ_2012_001', '4', path.joinpath('4.오산시_연령별_거주인구격자(총인구).geojson'))
    GetCompasData('SBJ_2012_001', '5', path.joinpath('5.오산시_연령별_거주인구격자(유소년).geojson'))
    GetCompasData('SBJ_2012_001', '6', path.joinpath('6.오산시_연령별_거주인구격자(생산가능인구).geojson'))
    GetCompasData('SBJ_2012_001', '7', path.joinpath('7.오산시_연령별_거주인구격자(고령).geojson'))
    GetCompasData('SBJ_2012_001', '8', path.joinpath('8.오산시_유동인구(2019).csv'))
    GetCompasData('SBJ_2012_001', '9', path.joinpath('9.오산시_어린이보호구역.csv'))
    GetCompasData('SBJ_2012_001', '10', path.joinpath('10.오산시_학교위치정보.csv'))
    GetCompasData('SBJ_2012_001', '11', path.joinpath('11.오산시_초등학교_통학구.geojson'))
    GetCompasData('SBJ_2012_001', '12', path.joinpath('12.오산시_중학교_학군.geojson'))
    GetCompasData('SBJ_2012_001', '13', path.joinpath('13.오산시_어린이집_유치원현황.csv'))
    GetCompasData('SBJ_2012_001', '14', path.joinpath('14.오산시_기상데이터(2010~2019).csv'))
    GetCompasData('SBJ_2012_001', '15', path.joinpath('15.오산시_무인교통단속카메라.csv'))
    GetCompasData('SBJ_2012_001', '16', path.joinpath('16.오산시_도로안전표지표준데이터.csv'))
    GetCompasData('SBJ_2012_001', '17', path.joinpath('17.오산시_횡단보도.geojson'))
    GetCompasData('SBJ_2012_001', '18', path.joinpath('18.오산시_과속방지턱표준데이터.csv'))
    GetCompasData('SBJ_2012_001', '19', path.joinpath('19.오산시_신호등.geojson'))
    GetCompasData('SBJ_2012_001', '20', path.joinpath('20.오산시_CCTV설치현황.csv'))
    GetCompasData('SBJ_2012_001', '21', path.joinpath('21.오산시_인도.geojson'))
    GetCompasData('SBJ_2012_001', '22', path.joinpath('22.오산시_버스정류장.csv'))
    GetCompasData('SBJ_2012_001', '23', path.joinpath('23.오산시_상세도로망_LV6.geojson'))
    GetCompasData('SBJ_2012_001', '24', path.joinpath('24.평일_전일,시간대별_오산시_추정교통량_Level6.csv'))
    GetCompasData('SBJ_2012_001', '25', path.joinpath('25.평일_전일_오산시_혼잡빈도강도_Level6.csv'))
    GetCompasData('SBJ_2012_001', '26', path.joinpath('26.평일_전일_오산시_혼잡시간강도_Level6.csv'))
    GetCompasData('SBJ_2012_001', '27', path.joinpath('27.오산시_도로명주소_건물.geojson'))
    GetCompasData('SBJ_2012_001', '28', path.joinpath('28.오산시_건물연면적_격자.geojson'))
    GetCompasData('SBJ_2012_001', '29', path.joinpath('29.오산시_체육시설현황.csv'))
    GetCompasData('SBJ_2012_001', '30', path.joinpath('30.오산시_학원_및_교습소_현황.csv'))
    GetCompasData('SBJ_2012_001', '31', path.joinpath('31.오산시_법정경계(시군구).geojson'))
    GetCompasData('SBJ_2012_001', '32', path.joinpath('32.오산시_행정경계(읍면동).geojson'))
    GetCompasData('SBJ_2012_001', '33', path.joinpath('33.오산시_법정경계(읍면동).geojson'))
    GetCompasData('SBJ_2012_001', '34', path.joinpath('34.오산시_지적도.geojson'))
    GetCompasData('SBJ_2012_001', '35', path.joinpath('35.코드정의서.xlsx'))

> ## __Load Data__ 

주정차단속 = pd.read_csv(os.path.join(path,'1.오산시_주정차단속(2018~2020).csv'),encoding="utf-8")
어린이교통사고_격자 = gpd.read_file(os.path.join(path,'2.오산시_어린이교통사고_격자.geojson'))
차량등록현황_격자 = gpd.read_file(os.path.join(path,'3.오산시_차량등록현황_격자.geojson'))
거주인구격자_총인구 = gpd.read_file(os.path.join(path,'4.오산시_연령별_거주인구격자(총인구).geojson'))
거주인구격자_유소년 = gpd.read_file(os.path.join(path,'5.오산시_연령별_거주인구격자(유소년).geojson'))
거주인구격자_생산가능 = gpd.read_file(os.path.join(path,'6.오산시_연령별_거주인구격자(생산가능인구).geojson'))
거주인구격자_고령 = gpd.read_file(os.path.join(path,'7.오산시_연령별_거주인구격자(고령).geojson'))
유동인구 = pd.read_csv(os.path.join(path,'8.오산시_유동인구(2019).csv'),encoding="utf-8")
어린이보호구역 = pd.read_csv(os.path.join(path,'9.오산시_어린이보호구역.csv'),encoding="utf-8")
학교위치정보 = pd.read_csv(os.path.join(path,'10.오산시_학교위치정보.csv'),encoding="utf-8")
초등학교_통학구 = gpd.read_file(os.path.join(path,'11.오산시_초등학교_통학구.geojson'))
중학교_학군 = gpd.read_file(os.path.join(path,'12.오산시_중학교_학군.geojson'))
어린이집_유치원현황 = pd.read_csv(os.path.join(path,'13.오산시_어린이집_유치원현황.csv'),encoding="utf-8")
기상데이터 = pd.read_csv(os.path.join(path,'14.오산시_기상데이터(2010~2019).csv'),encoding="utf-8")
무인교통단속카메라 = pd.read_csv(os.path.join(path,'15.오산시_무인교통단속카메라.csv'),encoding="utf-8")
도로안전표지 = pd.read_csv(os.path.join(path,'16.오산시_도로안전표지표준데이터.csv'),encoding="utf-8")
횡단보도 = gpd.read_file(os.path.join(path,'17.오산시_횡단보도.geojson'))
과속방지턱 = pd.read_csv(os.path.join(path,'18.오산시_과속방지턱표준데이터.csv'),encoding="utf-8")
신호등 = gpd.read_file(os.path.join(path,'19.오산시_신호등.geojson'))
CCTV설치현황 = pd.read_csv(os.path.join(path,'20.오산시_CCTV설치현황.csv'),encoding="utf-8")
인도 = gpd.read_file(os.path.join(path,'21.오산시_인도.geojson'))
버스정류장 = pd.read_csv(os.path.join(path,'22.오산시_버스정류장.csv'),encoding="utf-8")
상세도로망 = gpd.read_file(os.path.join(path,'23.오산시_상세도로망_LV6.geojson'))
시간대별_추정교통량 = pd.read_csv(os.path.join(path,'24.평일_전일,시간대별_오산시_추정교통량_Level6.csv'),encoding="utf-8")
혼잡빈도강도 = pd.read_csv(os.path.join(path,'25.평일_전일_오산시_혼잡빈도강도_Level6.csv'),encoding="utf-8")
혼잡시간강도 = pd.read_csv(os.path.join(path,'26.평일_전일_오산시_혼잡시간강도_Level6.csv'),encoding="utf-8")
도로명주소_건물 = gpd.read_file(os.path.join(path,'27.오산시_도로명주소_건물.geojson'))
건물연면적_격자 = gpd.read_file(os.path.join(path,'28.오산시_건물연면적_격자.geojson'))
체육시설현황 = pd.read_csv(os.path.join(path,'29.오산시_체육시설현황.csv'),encoding="utf-8")
학원_및_교습소_현황 = pd.read_csv(os.path.join(path,'30.오산시_학원_및_교습소_현황.csv'),encoding="utf-8")
법정경계_시군구 = gpd.read_file(os.path.join(path,'31.오산시_법정경계(시군구).geojson'))
행정경계_읍면동 = gpd.read_file(os.path.join(path,'32.오산시_행정경계(읍면동).geojson'))
법정경계_읍면동 = gpd.read_file(os.path.join(path,'33.오산시_법정경계(읍면동).geojson'))
지적도 = gpd.read_file(os.path.join(path,'34.오산시_지적도.geojson'))

> ## __Preprocessing__  

> __1.주정차 단속 데이터 : 오입력 제거 및 결측치 대체__

# 행정구역 오입력 데이터 제거
print("원본 데이터:",주정차단속.shape[0])
to_drop = 주정차단속[주정차단속['행정구역'].str.contains('오산시')==False]
주정차단속.drop(to_drop.index,inplace=True)
print("제거 후 데이터:",주정차단속.shape[0])
print("오입력 데이터 예시,")
to_drop.head(1)

# 결측 데이터 조사 및 단속 위치 확인
print("각 칼럼 별 결측 비율(%) :")
print(주정차단속.isnull().sum()*100/주정차단속.shape[0])
print("위/경도 결측 데이터 단속 위치",주정차단속[주정차단속["단속위치_경도"].isnull()]["단속위치"].unique())

[구글맵 : 오산시 유탑유블레스](https://www.google.co.kr/maps/search/%EA%B2%BD%EA%B8%B0%EB%8F%84+%EC%98%A4%EC%82%B0%EC%8B%9C+%EC%9B%90%EB%8F%99+360-15+%EB%B2%88%EC%A7%80+%EC%99%B8+3%ED%95%84%EC%A7%80/@37.1414858,127.0701808,17z/data=!3m1!4b1?hl=ko)  

# 구글맵 기반 오산시 유탑유블레스 남측 위/경도 입력
주정차단속["단속위치_경도"].fillna(127.0701808,inplace=True)
주정차단속["단속위치_위도"].fillna(37.1414858,inplace=True)
주정차단속.isnull().sum()

> __2. ~ 3. : 결측 데이터 없음__  
(어린이 교통사고 데이터 , 차량등록 현황)

> __4.~7.연령별 거주인구 격자데이터 (총인구, 유소년인구, 생산가능인구, 고령인구) : 결측치 대체__

# 각 정보별 결측 데이터 비율
print("총 인구 데이터 결측비율 :",거주인구격자_총인구.isnull().sum().sum()/거주인구격자_총인구.shape[0])
print("유소년 인구 데이터 결측비율 :",거주인구격자_유소년.isnull().sum().sum()/거주인구격자_유소년.shape[0])
print("생산가능 인구 데이터 결측비율 :",거주인구격자_생산가능.isnull().sum().sum()/거주인구격자_생산가능.shape[0])
print("고령 인구 데이터 결측비율 :",거주인구격자_고령.isnull().sum().sum()/거주인구격자_고령.shape[0])

# 각 정보 결측 데이터 성격 : 격자내의 총 인구수
# 0명으로 결측값 처리

거주인구격자_총인구.fillna(0,inplace=True)
거주인구격자_유소년.fillna(0,inplace=True)
거주인구격자_생산가능.fillna(0,inplace=True)
거주인구격자_고령.fillna(0,inplace=True)

print("총 인구 데이터 결측비율 :",거주인구격자_총인구.isnull().sum().sum()/거주인구격자_총인구.shape[0])
print("유소년 인구 데이터 결측비율 :",거주인구격자_유소년.isnull().sum().sum()/거주인구격자_유소년.shape[0])
print("생산가능 인구 데이터 결측비율 :",거주인구격자_생산가능.isnull().sum().sum()/거주인구격자_생산가능.shape[0])
print("고령 인구 데이터 결측비율 :",거주인구격자_고령.isnull().sum().sum()/거주인구격자_고령.shape[0])

> __8.유동인구 : 관심 데이터 재구조 및 별도 저장__  

# 등*하교 시간대를 관심시간대로 설정하여 월/시간별 데이터 전체의 평균 유동인구 계산

관심시간대 = ["TMST_08","TMST_09","TMST_10","TMST_12","TMST_13","TMST_14","TMST_15","TMST_16","TMST_17","TMST_18"]

평균유동인구 = 유동인구.drop(["STD_YM"],axis=1).groupby(['lon','lat']).mean()[관심시간대].mean(axis=1)
평균유동인구 = 평균유동인구.reset_index().rename(columns={0:"평균유동인구"})
평균유동인구.head(3)

> __9.어린이보호구역 현황 : 오입력 데이터 제거__  

# CCTV 설치 여부 정보와 CCTV 설치대수 정보가 서로 모순인 행 존재 : 제거
condition1 = ((어린이보호구역["CCTV설치여부"]=="Y")&(어린이보호구역["CCTV설치대수"]==0))
condition2 = ((어린이보호구역["CCTV설치여부"]=="N")&(어린이보호구역["CCTV설치대수"]!=0))
어린이보호구역[condition1|condition2]

어린이보호구역.drop([23],inplace=True)

> __10.학교위치정보 : 관심 데이터 추출 및 별도 저장__

# 관심 학교위치 정보 추출
print("기존 데이터:",학교위치정보.shape[0])
초등학교위치정보 = 학교위치정보[학교위치정보.학교구분=="초등학교"]
print("전처리 후 데이터:",초등학교위치정보.shape[0])

> __11. ~ 13. : 결측치 없음__  
(초등학교 통학구 , 중학교 학군 , 어린이집_유치원 현황)

> __14.기상데이터 : 분석과정에서 미사용__

> __15. ~ 19. : 결측치 없음__  
(무인교통단속카메라 , 도로안전표지 , 횡단보도 , 과속방지턱 , 신호등)

> __20.CCTV 설치 현황 : 결측치 제거__

print("각 칼럼 별 결측 비율(%) :")
CCTV설치현황.isnull().sum()*100/CCTV설치현황.shape[0]

# 설치위치_도로명주소 결측치 : "미기재" 로 대체
CCTV설치현황["설치위치_도로명주소"].fillna("미기재",inplace=True)

print("각 칼럼 별 결측 비율(%) :")
CCTV설치현황.isnull().sum()*100/CCTV설치현황.shape[0]

> __21. ~ 22. : 결측치 없음__  
(인도 , 버스정류장)

> __23.상세 도로망 데이터__

print("각 칼럼 별 결측 비율(%) :")
상세도로망.isnull().sum()*100/상세도로망.shape[0]

# 결측을 제외한 정보량이 크게 부족한 facil_name 의 경우 drop , road_name 의 경우 "미기재" 로 대체

상세도로망.drop(["facil_name"],axis=1,inplace=True)
상세도로망["road_name"].fillna("미기재",inplace=True)

print("전처리 후 각 칼럼 별 결측 비율(%) :")
상세도로망.isnull().sum()*100/상세도로망.shape[0]

> __24.시간대별 추정 교통량 데이터__

print("각 칼럼별 결측 데이터 비율(%) :")
시간대별_추정교통량.isnull().sum()*100/시간대별_추정교통량.shape[0]

# 도로명 결측 데이터 : "미기재" 로 대체
시간대별_추정교통량["도로명"].fillna("미기재",inplace=True)

print("전처리 후 각 칼럼별 결측 데이터 비율(%) :")
시간대별_추정교통량.isnull().sum()*100/시간대별_추정교통량.shape[0]

# 시간 데이터에 데이터타입 혼용으로 float 타입으로 통일
print("원본 시간 칼럼 데이터 타입:",시간대별_추정교통량["시간적범위"][시간대별_추정교통량["시간적범위"]!="전일"].map(type).unique())
    
시간대별_추정교통량["시간적범위"][시간대별_추정교통량["시간적범위"]!="전일"]=시간대별_추정교통량["시간적범위"][시간대별_추정교통량["시간적범위"]!="전일"].map(np.int32)
print("수정 후 시간 칼럼 데이터 타입:",시간대별_추정교통량["시간적범위"][시간대별_추정교통량["시간적범위"]!="전일"].map(type).unique())
시간대별_추정교통량["시간적범위"].value_counts()

# 상세도로망_LinkID 칼럼으로부터 상하행 정보 추출 및 별도 칼럼 저장
시간대별_추정교통량['상하행'] = 시간대별_추정교통량['상세도로망_LinkID'].map(lambda x:str(x)[-2:]).replace({'01':'상행','02':'하행'})
시간대별_추정교통량["상세도로망_LinkID"]=시간대별_추정교통량["상세도로망_LinkID"].map(lambda x:str(x)[:-2])
시간대별_추정교통량.head(3)

> __25. ~ 26. 혼잡 강도 데이터 (빈도 , 시간)__

print("각 칼럼별 결측 데이터 비율(%) :")
혼잡빈도강도.isnull().sum()*100/혼잡빈도강도.shape[0],혼잡시간강도.isnull().sum()*100/혼잡시간강도.shape[0]

# 동일한 도로에 대해 도로명이 결측 : "미기재" 로 대체

common = set(혼잡빈도강도[혼잡빈도강도["도로명"].isnull()]["상세도로망_LinkID"].values)-set(혼잡시간강도[혼잡시간강도["도로명"].isnull()]["상세도로망_LinkID"].values)

same_nan_idx = True if not common else False
print("동일한 도로에 대해 결측인가 :",same_nan_idx)

혼잡빈도강도["도로명"].fillna("미기재",inplace=True)
혼잡시간강도["도로명"].fillna("미기재",inplace=True)

print("전처리 후 각 칼럼별 결측 데이터 비율(%) :")
혼잡빈도강도.isnull().sum()*100/혼잡빈도강도.shape[0] , 혼잡시간강도.isnull().sum()*100/혼잡시간강도.shape[0]

> __27.도로명주소 건물 데이터__

print("각 칼럼별 결측 데이터 비율(%) :")
도로명주소_건물.isnull().sum()*100/도로명주소_건물.shape[0]

# 두 칼럼 모두 결측인 경우 "미기재" 로 대체 , 상세건물명(BULD_NM_DC)만 결측인 경우 건물명(BULD_NM)으로 대체

도로명주소_건물["BULD_NM"].fillna("미기재",inplace=True)
도로명주소_건물["BULD_NM_DC"].fillna(도로명주소_건물["BULD_NM"],inplace=True)

print("전처리 후 각 칼럼별 결측 데이터 비율(%) :")
도로명주소_건물.isnull().sum()*100/도로명주소_건물.shape[0]

> __28.건물연면적 격자 데이터 : 분석과정에서 미사용__

> __30.학원 및 교습소 현황 데이터__

print("각 칼럼별 결측 데이터 비율(%) :")
학원_및_교습소_현황.isnull().sum()*100/학원_및_교습소_현황.shape[0]

# 교습과정명 결측 데이터 제거 및 관심 데이터 선정

학원_및_교습소_현황.dropna(inplace=True)

학원_및_교습소_현황 = 학원_및_교습소_현황[학원_및_교습소_현황["업종구분명"]=="학교교과교습학원"] # 평생직업교육학원 및 교습소 제거        

drop_list = ["중등","고등","대학","일반","성인","입시"] # 필터링 단어 선정

def drop_check(name):
    if any([True for element in drop_list if element in name]):
        return True
    return False

drop_idx = 학원_및_교습소_현황['교습과정명'].map(drop_check)

관심학원_현황 = 학원_및_교습소_현황[drop_idx!=True]

print("전처리 후 각 칼럼별 결측 데이터 비율(%) :")
학원_및_교습소_현황.isnull().sum()*100/학원_및_교습소_현황.shape[0]

> __29. , 31. , 32. , 33. , 34. :결측치 없음__  
(체육시설 현황 , 법정경계_시군구 , 행정경계_읍면동 , 법정경계_읍면동 , 지적도)

> ## __Save preprocessed data__  

주정차단속.to_csv(os.path.join(path,'1.오산시_주정차단속(2018~2020).csv'),encoding="utf-8",index=False)
어린이교통사고_격자.to_file(os.path.join(path,'2.오산시_어린이교통사고_격자.geojson'),driver='GeoJSON')
차량등록현황_격자.to_file(os.path.join(path,'3.오산시_차량등록현황_격자.geojson'),driver='GeoJSON')
거주인구격자_총인구.to_file(os.path.join(path,'4.오산시_연령별_거주인구격자(총인구).geojson'),driver='GeoJSON')
거주인구격자_유소년.to_file(os.path.join(path,'5.오산시_연령별_거주인구격자(유소년).geojson'),driver='GeoJSON')
거주인구격자_생산가능.to_file(os.path.join(path,'6.오산시_연령별_거주인구격자(생산가능인구).geojson'),driver='GeoJSON')
거주인구격자_고령.to_file(os.path.join(path,'7.오산시_연령별_거주인구격자(고령).geojson'),driver='GeoJSON')
평균유동인구.to_csv(os.path.join(path,'평균유동인구.csv'),encoding="utf-8",index=False)
어린이보호구역.to_csv(os.path.join(path,'9.오산시_어린이보호구역.csv'),encoding="utf-8",index=False)
초등학교위치정보.to_csv(os.path.join(path,'초등학교위치정보.csv'),encoding="utf-8",index=False)
초등학교_통학구.to_file(os.path.join(path,'11.오산시_초등학교_통학구.geojson'),driver='GeoJSON')
중학교_학군.to_file(os.path.join(path,'12.오산시_중학교_학군.geojson'),driver='GeoJSON')
어린이집_유치원현황.to_csv(os.path.join(path,'13.오산시_어린이집_유치원현황.csv'),encoding="utf-8",index=False)
기상데이터.to_csv(os.path.join(path,'14.오산시_기상데이터(2010~2019).csv'),encoding="utf-8",index=False)
무인교통단속카메라.to_csv(os.path.join(path,'15.오산시_무인교통단속카메라.csv'),encoding="utf-8",index=False)
도로안전표지.to_csv(os.path.join(path,'16.오산시_도로안전표지표준데이터.csv'),encoding="utf-8",index=False)
횡단보도.to_file(os.path.join(path,'17.오산시_횡단보도.geojson'),driver='GeoJSON')
과속방지턱.to_csv(os.path.join(path,'18.오산시_과속방지턱표준데이터.csv'),encoding="utf-8",index=False)
신호등.to_file(os.path.join(path,'19.오산시_신호등.geojson'),driver='GeoJSON')
CCTV설치현황.to_csv(os.path.join(path,'20.오산시_CCTV설치현황.csv'),encoding="utf-8",index=False)
인도.to_file(os.path.join(path,'21.오산시_인도.geojson'),driver='GeoJSON')
버스정류장.to_csv(os.path.join(path,'22.오산시_버스정류장.csv'),encoding="utf-8",index=False)
상세도로망.to_file(os.path.join(path,'23.오산시_상세도로망_LV6.geojson'),driver='GeoJSON')
시간대별_추정교통량.to_csv(os.path.join(path,'24.평일_전일,시간대별_오산시_추정교통량_Level6.csv'),encoding="utf-8",index=False)
혼잡빈도강도.to_csv(os.path.join(path,'25.평일_전일_오산시_혼잡빈도강도_Level6.csv'),encoding="utf-8",index=False)
혼잡시간강도.to_csv(os.path.join(path,'26.평일_전일_오산시_혼잡시간강도_Level6.csv'),encoding="utf-8",index=False)
도로명주소_건물.to_file(os.path.join(path,'27.오산시_도로명주소_건물.geojson'),driver='GeoJSON')
건물연면적_격자.to_file(os.path.join(path,'28.오산시_건물연면적_격자.geojson'),driver='GeoJSON')
체육시설현황.to_csv(os.path.join(path,'29.오산시_체육시설현황.csv'),encoding="utf-8",index=False)
학원_및_교습소_현황.to_csv(os.path.join(path,'30.오산시_학원_및_교습소_현황.csv'),encoding="utf-8",index=False)
관심학원_현황.to_csv(os.path.join(path,'관심학원_현황.csv'),encoding="utf-8",index=False)
법정경계_시군구.to_file(os.path.join(path,'31.오산시_법정경계(시군구).geojson'),driver='GeoJSON')
행정경계_읍면동.to_file(os.path.join(path,'32.오산시_행정경계(읍면동).geojson'),driver='GeoJSON')
법정경계_읍면동.to_file(os.path.join(path,'33.오산시_법정경계(읍면동).geojson'),driver='GeoJSON')
지적도.to_file(os.path.join(path,'34.오산시_지적도.geojson'),driver='GeoJSON')
